pyproject.toml
[tool.poetry]
name = "fraud-detection-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
gqlalchemy = {path = "./gqlalchemy", develop = true}
pyarrow = "^7.0.0"
python-dotenv = "^0.19.2"
ipykernel = "^6.9.0"
pandas = "^1.4.1"
scikit-learn = "^1.0.2"
matplotlib = "^3.5.1"
seaborn = "^0.11.2"
PyYAML = "^6.0"
notebook = "^6.4.8"
importlib-resources = "^5.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
